services/function-deployment/requirements.txt
asyncpg
kafka-python
kubernetes